mesh and particle dynamics share the same base
